241112 03:47 황진 - 자유게시판 목록, 상세, 작성, 수정, 삭제작업

+++ src/main/java/com/example/moamz/service/community/free/FreeBoardService.java
@@ -1,6 +1,7 @@
 package com.example.moamz.service.community.free;
 
 import com.example.moamz.domain.dto.community.free.*;
+import com.example.moamz.mapper.community.PostMapper;
 import com.example.moamz.mapper.community.free.FreeLikeMapper;
 import com.example.moamz.mapper.community.free.FreeMapper;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +18,7 @@ import java.util.List;
 public class FreeBoardService {
     public final FreeMapper freeMapper;
     public final FreeLikeMapper freeLikeMapper;
+    private final PostMapper postMapper;
 
     // 자유게시판 글 등록 메서드
     public void registFree(FreeWriteDTO freeWriteDTO) {
@@ -31,8 +33,18 @@ public class FreeBoardService {
     }
 
     // 자유게시판 글 삭제 메서드
-    public void removeFree(Long postId) {
-        freeMapper.deleteFree(postId);
+    public boolean removeFree(Long userCode, Long postId) {
+
+        // 게시글 작성자
+        Long writerCode = postMapper.selectWriterCode(postId);
+
+        // 삭제를 요청한 userCode와 게시글 작성자 userCode가 일치할 때만 삭제한다.
+        if(userCode.equals(writerCode)) {
+            freeMapper.deleteFree(postId);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     // 자유게시판 목록 메서드
@@ -40,8 +52,17 @@ public class FreeBoardService {
         return freeMapper.selectFreeList();
     }
 
-    // 자유게시판 상세 메서드
-    public FreeDetailDTO findFreeDetail(Long postId) {
+    // 자유게시판 상세보기 메서드
+    public FreeDetailDTO findFreeDetail(Long postId, Long userCode) {
+
+        // 작성자 userCode
+        Long writerCode = postMapper.selectWriterCode(postId);
+
+        // 작성자 Code랑 세션의 userCode 다를때 조회수 +1
+        if(!writerCode.equals(userCode)) {
+            postMapper.updateViewCount(postId);
+        }
+
         return freeMapper.selectFreeDetail(postId)
                 .orElseThrow(() -> new IllegalStateException("❌❌❌유효하지 않은 게시글입니다."));
     }
